Update builddb_rest API implementation
- Utilize newest cbbuild-* libs
- Utilize couchbase SDK 4.x
- Convert to uv project
- Make Docker container run as non-root

Reviewed-on: https://review.couchbase.org/c/build-infra/+/226681
Tested-by: Chris Hillery <ceej@couchbase.com>

+++ builddb_rest/scripts/builddb_rest_start.sh
@@ -1,14 +1,23 @@
 #!/usr/bin/env bash
 
-# Ensure an up-to-date (unmodified) product-metadata checkout
-# is available
-cd /home/couchbase
-rm -rf product-metadata
-git clone https://github.com/couchbase/product-metadata > /dev/null
 
 # Start pserve (as new, long-running, foreground process)
 [[ "$1" == "default" ]] && {
-    exec /usr/local/bin/pserve /etc/builddb_rest.ini
+    # Ensure the builddb_rest.ini file is present
+    if [ ! -f /etc/builddb_rest.ini ]; then
+        echo "Error: /etc/builddb_rest.ini not found."
+        exit 1
+    fi
+
+    # Ensure an up-to-date (unmodified) product-metadata checkout
+    # is available
+    cd /home/couchbase
+    rm -rf product-metadata
+    git clone https://github.com/couchbase/product-metadata > /dev/null
+
+    # Start the pserve server
+    cd /home/couchbase/builddb_rest
+    exec uv run pserve /etc/builddb_rest.ini
 }
 
 exec "$@"
